+++ src/main/java/org/mmarini/wheelly/swing/UIController.java
@@ -32,7 +32,6 @@ package org.mmarini.wheelly.swing;
 import com.fasterxml.jackson.databind.JsonNode;
 import hu.akarnokd.rxjava3.swing.SwingObservable;
 import io.reactivex.Flowable;
-import io.reactivex.processors.BehaviorProcessor;
 import io.reactivex.rxjava3.core.BackpressureStrategy;
 import org.mmarini.Tuple2;
 import org.mmarini.wheelly.model.*;
@@ -65,29 +64,11 @@ public class UIController {
     private static final String CONFIG_FILE = ".wheelly.yml";
     private static final Logger logger = LoggerFactory.getLogger(UIController.class);
 
-    private static final long REMOTE_CLOCK_PERIOD = 60;   // Seconds
-    private static final int REMOTE_CLOCK_SAMPLES = 10;
-    private static final long STATUS_INTERVAL = 200; // Millis
-    private static final long MOTOR_PERIOD = 100;
-    private static final long MOTOR_INTERVAL = 100;
-
     /**
      *
      */
     public static UIController create() {
         return new UIController();
-    }
-
-    /**
-     * @param x
-     * @param y
-     */
-    static MotorCommand.Direction toDir(float x, float y) {
-        double dir = atan2(y, x) + Math.PI;
-        double deg = toDegrees(dir);
-        int i = ((int) round(dir * 8 / Math.PI)) % 16;
-        int d = ((i + 13) / 2) % 8;
-        return MotorCommand.Direction.values()[d + 1];
     }
 
     /**
@@ -105,38 +86,13 @@ public class UIController {
                 .isPresent();
     }
 
-    /**
-     * @param motorCommands
-     */
-    private static boolean isPassingCommand(List<MotorCommand> motorCommands) {
-        MotorCommand m0 = motorCommands.get(0);
-        MotorCommand m1 = motorCommands.get(1);
-        return m0.isRunning() || m1.isRunning();
-    }
-
-    /**
-     * @param xy
-     */
-    private static MotorCommand toMotorCommand(Tuple2<Float, Float> xy) {
-        float x = xy._1;
-        float y = xy._2;
-        float speed = (float) sqrt(x * x + y * y);
-        MotorCommand.Direction direction = MotorCommand.Direction.NONE;
-        if (speed > 0.1f) {
-            direction = toDir(x, y);
-        }
-        return MotorCommand.empty().setDirection(direction).setSpeed(speed);
-    }
-
     private final PreferencesPane preferencesPane;
     private final WheellyFrame frame;
     private final Dashboard dashboard;
     private final Radar radar;
-    private RxJoystick joystick;
-    private RxController wheellyController;
     private String baseUrl;
     private String joystickPort;
-    private BehaviorProcessor<RemoteClock> remoteClock;
+    private FlowBuilder flowBuilder;
 
     /**
      *
@@ -160,73 +116,6 @@ public class UIController {
     }
 
     /**
-     * Handles the application on window open.
-     * Opens current configuration
-     */
-    private void handleWindowOpened() {
-        openConfig();
-    }
-
-    /**
-     * Returns the UIController with current configuration opened
-     */
-    private void openConfig() {
-        this.wheellyController = RxController.create(baseUrl);
-        this.joystick = RxJoystick.create(joystickPort);
-
-        this.remoteClock = BehaviorProcessor.create();
-        wheellyController.remoteClock(REMOTE_CLOCK_PERIOD, TimeUnit.SECONDS, REMOTE_CLOCK_SAMPLES)
-                .subscribe(remoteClock);
-
-        Flowable<StatusBody> statusFlow = Flowable.interval(0, STATUS_INTERVAL, TimeUnit.MILLISECONDS)
-                .flatMap(x -> wheellyController.status());
-        Flowable<StatusBody> scanCommand = Flowable.mergeArray(
-                        joystick.getValues(RxJoystick.BUTTON_0),
-                        joystick.getValues(RxJoystick.BUTTON_1),
-                        joystick.getValues(RxJoystick.BUTTON_2),
-                        joystick.getValues(RxJoystick.BUTTON_3))
-                .filter(x -> x > 0f)
-                .flatMap(x -> wheellyController.scan());
-
-        Flowable<MotorCommand> joystickCommand = joystick.getXY()
-                .map(UIController::toMotorCommand);
-        Flowable<StatusBody> motorCommand = Flowable.interval(0, MOTOR_PERIOD, TimeUnit.MILLISECONDS)
-                .withLatestFrom(joystickCommand, (t, cmd) -> cmd.setTick(t))
-                .buffer(2, 1)
-                .filter(UIController::isPassingCommand)
-                .map(x -> x.get(1))
-                .withLatestFrom(remoteClock, (cmd, clock) -> cmd.setClock(clock))
-                .flatMap(cmd ->
-                        wheellyController.moveTo(
-                                cmd.getLeftSpeed(),
-                                cmd.getRightSpeed(),
-                                cmd.clock.toRemote(Instant.now().plusMillis(MOTOR_INTERVAL))));
-
-        Flowable.merge(statusFlow, scanCommand, motorCommand)
-                .withLatestFrom(remoteClock, RxController::toStatus)
-                .doOnNext(s -> {
-                    dashboard.setPower(s.voltage.value);
-                    dashboard.setObstacleDistance(
-                            getValue(s.obstacles, 90)
-                                    .map(InstantValue::get)
-                                    .orElse(400));
-                    dashboard.setMotors(s.direction.value._1, s.direction.value._2);
-                    dashboard.setForwardBlock(isBlockDistance(s));
-                    radar.setSamples(s.obstacles);
-                })
-                .subscribe();
-    }
-
-    /**
-     * Returns the UIController with current configuration closed
-     */
-    private UIController closeConfig() {
-        remoteClock.onComplete();
-        joystick.close();
-        return this;
-    }
-
-    /**
      *
      */
     private UIController buildFlow() {
@@ -239,24 +128,11 @@ public class UIController {
     }
 
     /**
-     * @throws IOException in case of error
+     * Returns the UIController with current configuration closed
      */
-    private void loadConfig() throws IOException {
-        JsonNode node = fromFile(CONFIG_FILE);
-        Yaml.config().apply(Locator.root()).accept(node);
-        joystickPort = Yaml.joystickPort(node);
-        baseUrl = Yaml.baseUrl(node);
-    }
-
-    /**
-     *
-     */
-    private void saveConfig() {
-        try {
-            Yaml.saveConfig(CONFIG_FILE, joystickPort, baseUrl);
-        } catch (IOException ex) {
-            logger.error(ex.getMessage(), ex);
-        }
+    private UIController closeConfig() {
+        flowBuilder.detach();
+        return this;
     }
 
     /**
@@ -272,6 +148,53 @@ public class UIController {
             joystickPort = preferencesPane.getJoystickPort();
             baseUrl = preferencesPane.getBaseUrl();
             saveConfig();
+        }
+    }
+
+    /**
+     * Handles the application on window open.
+     * Opens current configuration
+     */
+    private void handleWindowOpened() {
+        openConfig();
+    }
+
+    /**
+     * @throws IOException in case of error
+     */
+    private void loadConfig() throws IOException {
+        JsonNode node = fromFile(CONFIG_FILE);
+        Yaml.config().apply(Locator.root()).accept(node);
+        joystickPort = Yaml.joystickPort(node);
+        baseUrl = Yaml.baseUrl(node);
+    }
+
+    /**
+     * Returns the UIController with current configuration opened
+     */
+    private void openConfig() {
+        this.flowBuilder = FlowBuilder.create(RxController.create(baseUrl), RxJoystick.create(joystickPort));
+        flowBuilder.getStatus()
+                .doOnNext(s -> {
+                    dashboard.setPower(s.voltage.value);
+                    dashboard.setObstacleDistance(
+                            getValue(s.obstacles, 90)
+                                    .map(InstantValue::get)
+                                    .orElse(400));
+                    dashboard.setMotors(s.direction.value._1, s.direction.value._2);
+                    dashboard.setForwardBlock(isBlockDistance(s));
+                    radar.setSamples(s.obstacles);
+                }).subscribe();
+    }
+
+    /**
+     *
+     */
+    private void saveConfig() {
+        try {
+            Yaml.saveConfig(CONFIG_FILE, joystickPort, baseUrl);
+        } catch (IOException ex) {
+            logger.error(ex.getMessage(), ex);
         }
     }