+++ CHANGELOG.md
@@ -9,4 +9,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Added
 
-- Issue #3: Asynchronous arduino classes
+- Issue #2: Manage motor pulse in arduino
+- Issue #3: Add Led image